Grim Fandango Remastered — Syntax update
* Update script to 2.11 syntax
* Update authors list
* Use generic dependencies
* Drop call to postinst_icons_linking

+++ play.it-2/games/play-grim-fandango-remastered.sh
@@ -3,6 +3,7 @@ set -o errexit
 
 ###
 # Copyright (c) 2015-2019, Antoine "vv221/vv222" Le Gonidec
+# Copyright (c) 2016-2019, Solène "Mopi" Huault
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -30,18 +31,16 @@ set -o errexit
 
 ###
 # Grim Fandango Remastered
-# build native Linux packages from the original installers
+# build native packages from the original installers
 # send your bug reports to vv221@dotslashplay.it
 ###
 
-script_version=20180224.1
+script_version=20191130.1
 
 # Set game-specific variables
 
 GAME_ID='grim-fandango'
 GAME_NAME='Grim Fandango Remastered'
-
-ARCHIVES_LIST='ARCHIVE_GOG'
 
 ARCHIVE_GOG='gog_grim_fandango_remastered_2.3.0.7.sh'
 ARCHIVE_GOG_URL='https://www.gog.com/game/grim_fandango_remastered'
@@ -50,26 +49,24 @@ ARCHIVE_GOG_SIZE='6100000'
 ARCHIVE_GOG_VERSION='1.4-gog2.3.0.7'
 ARCHIVE_GOG_TYPE='mojosetup_unzip'
 
-ARCHIVE_DOC1_PATH='data/noarch/docs'
-ARCHIVE_DOC1_FILES='./*'
+ARCHIVE_DOC0_DATA_PATH='data/noarch/docs'
+ARCHIVE_DOC0_DATA_FILES='*'
 
-ARCHIVE_DOC2_PATH='data/noarch/game/bin'
-ARCHIVE_DOC2_FILES='./*License.txt ./common-licenses'
+ARCHIVE_DOC1_DATA_PATH='data/noarch/game/bin'
+ARCHIVE_DOC1_DATA_FILES='*License.txt common-licenses'
 
 ARCHIVE_GAME_BIN_PATH='data/noarch/game/bin'
-ARCHIVE_GAME_BIN_FILES='./GrimFandango ./*.so ./libSDL2-2.0.so.1 ./x86'
+ARCHIVE_GAME_BIN_FILES='GrimFandango *.so libSDL2-2.0.so.1 x86'
 
 ARCHIVE_GAME_MOVIES_PATH='data/noarch/game/bin'
-ARCHIVE_GAME_MOVIES_FILES='./MoviesHD'
+ARCHIVE_GAME_MOVIES_FILES='MoviesHD'
 
 ARCHIVE_GAME_DATA_PATH='data/noarch/game/bin'
-ARCHIVE_GAME_DATA_FILES='./*.lab ./*.LAB ./controllerdef.txt ./en_gagl088.lip ./FontsHD ./*.tab ./icon.png ./patch_v2_or_v3_to_v4.bin ./patch_v4_to_v5.bin'
+ARCHIVE_GAME_DATA_FILES='*.lab *.LAB controllerdef.txt en_gagl088.lip FontsHD *.tab icon.png patch_v2_or_v3_to_v4.bin patch_v4_to_v5.bin'
 
 APP_MAIN_TYPE='native'
 APP_MAIN_EXE='GrimFandango'
-APP_MAIN_ICONS_LIST='APP_MAIN_ICON'
 APP_MAIN_ICON='icon.png'
-APP_MAIN_ICON_RES='128'
 
 PACKAGES_LIST='PKG_DATA PKG_MOVIES PKG_BIN'
 
@@ -80,55 +77,55 @@ PKG_MOVIES_ID="${GAME_ID}-movies"
 PKG_MOVIES_DESCRIPTION='movies'
 
 PKG_BIN_ARCH='32'
-PKG_BIN_DEPS_DEB="$PKG_MOVIES_ID, $PKG_DATA_ID, libc6, libstdc++6, libglu1-mesa | libglu1, libsdl2-2.0-0"
-PKG_BIN_DEPS_ARCH="$PKG_MOVIES_ID $PKG_DATA_ID lib32-glu lib32-sdl2"
+PKG_BIN_DEPS="$PKG_MOVIES_ID $PKG_DATA_ID glibc libstdc++ glu sdl2"
 
 # Load common functions
 
-target_version='2.3'
+target_version='2.11'
 
 if [ -z "$PLAYIT_LIB2" ]; then
-	[ -n "$XDG_DATA_HOME" ] || XDG_DATA_HOME="$HOME/.local/share"
-	if [ -e "$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh" ]; then
-		PLAYIT_LIB2="$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh"
-	elif [ -e './libplayit2.sh' ]; then
-		PLAYIT_LIB2='./libplayit2.sh'
-	else
-		printf '\n\033[1;31mError:\033[0m\n'
-		printf 'libplayit2.sh not found.\n'
-		exit 1
-	fi
+	: "${XDG_DATA_HOME:="$HOME/.local/share"}"
+	for path in\
+		"$PWD"\
+		"$XDG_DATA_HOME/play.it"\
+		'/usr/local/share/games/play.it'\
+		'/usr/local/share/play.it'\
+		'/usr/share/games/play.it'\
+		'/usr/share/play.it'
+	do
+		if [ -e "$path/libplayit2.sh" ]; then
+			PLAYIT_LIB2="$path/libplayit2.sh"
+			break
+		fi
+	done
 fi
-#shellcheck source=play.it-2/lib/libplayit2.sh
+if [ -z "$PLAYIT_LIB2" ]; then
+	printf '\n\033[1;31mError:\033[0m\n'
+	printf 'libplayit2.sh not found.\n'
+	exit 1
+fi
+# shellcheck source=play.it-2/lib/libplayit2.sh
 . "$PLAYIT_LIB2"
 
 # Extract game data
 
 extract_data_from "$SOURCE_ARCHIVE"
+prepare_package_layout
+rm --recursive "$PLAYIT_WORKDIR/gamedata"
 
-PKG='PKG_BIN'
-organize_data 'GAME_BIN' "$PATH_GAME"
-
-PKG='PKG_MOVIES'
-organize_data 'GAME_MOVIES' "$PATH_GAME"
+# Get icon
 
 PKG='PKG_DATA'
-organize_data 'DOC1'      "$PATH_DOC"
-organize_data 'DOC2'      "$PATH_DOC"
-organize_data 'GAME_DATA' "$PATH_GAME"
-
-rm --recursive "$PLAYIT_WORKDIR/gamedata"
+icons_get_from_package 'APP_MAIN'
 
 # Write launchers
 
 PKG='PKG_BIN'
-write_launcher 'APP_MAIN'
+launchers_write 'APP_MAIN'
 
 # Build package
 
-postinst_icons_linking 'APP_MAIN'
-write_metadata 'PKG_DATA'
-write_metadata 'PKG_MOVIES' 'PKG_BIN'
+write_metadata
 build_pkg
 
 # Clean up
